Rachel & Alyssa Work Session
Rewrote how the scheduleBuilder works and made the notifications function into a reusable factory that just needs to be passed in as a 'notify' dependancy in any controller you want to use it in. Schedule Builder is now just saving the schedules the librarian creates and displaying the latest schedule created.